requirements.txt
requests
python-dotenv
openai
python-dateutil
loguru
gradio
python-daemon
markdown2
schedule
